Fixed #5 and added constants
Pressing the space bar without the video selected should now not result in a scrolling anymore.

Plus, the script seems now more readable to me with named constants instead of random values. These constants will hopefully be used in the future to make the users able to decide what key to use for what action